Set marker colors for multi-effect plots

Convenience function to set theme marker colors for multi-effect forest plots. Effects without an explicit color will use these colors in order.

set_marker_colors(theme, colors)

Arguments

theme

A WebTheme object

colors

Character vector of colors for effects (e.g., c("#2563eb", "#dc2626"))

Value

Modified WebTheme object

Examples

web_theme_default() |>
  set_marker_colors(c("#0891b2", "#dc2626", "#16a34a"))
#> <webforest::WebTheme>
#>  @ name         : chr "default"
#>  @ colors       : <webforest::ColorPalette>
#>  .. @ background       : chr "#ffffff"
#>  .. @ foreground       : chr "#333333"
#>  .. @ primary          : chr "#0891b2"
#>  .. @ secondary        : chr "#64748b"
#>  .. @ accent           : chr "#8b5cf6"
#>  .. @ muted            : chr "#94a3b8"
#>  .. @ border           : chr "#e2e8f0"
#>  .. @ interval         : chr "#0891b2"
#>  .. @ interval_line    : chr "#475569"
#>  .. @ interval_positive: chr "#0891b2"
#>  .. @ interval_negative: chr "#dc2626"
#>  .. @ interval_neutral : chr "#64748b"
#>  .. @ summary_fill     : chr "#0891b2"
#>  .. @ summary_border   : chr "#0e7490"
#>  @ typography   : <webforest::Typography>
#>  .. @ font_family       : chr "system-ui, -apple-system, sans-serif"
#>  .. @ font_size_sm      : chr "0.75rem"
#>  .. @ font_size_base    : chr "0.875rem"
#>  .. @ font_size_lg      : chr "1rem"
#>  .. @ font_weight_normal: num 400
#>  .. @ font_weight_medium: num 500
#>  .. @ font_weight_bold  : num 600
#>  .. @ line_height       : num 1.5
#>  .. @ header_font_scale : num 1.05
#>  @ spacing      : <webforest::Spacing>
#>  .. @ row_height       : num 28
#>  .. @ header_height    : num 36
#>  .. @ section_gap      : num 16
#>  .. @ padding          : num 12
#>  .. @ container_padding: num 0
#>  .. @ axis_gap         : num 12
#>  .. @ group_padding    : num 8
#>  .. @ cell_padding_x   : num 10
#>  .. @ cell_padding_y   : num 4
#>  @ shapes       : <webforest::Shapes>
#>  .. @ point_size    : num 6
#>  .. @ summary_height: num 10
#>  .. @ line_width    : num 1.5
#>  .. @ border_radius : num 2
#>  .. @ marker_colors : chr [1:3] "#0891b2" "#dc2626" "#16a34a"
#>  .. @ marker_shapes : chr [1:4] "square" "circle" "diamond" "triangle"
#>  @ axis         : <webforest::AxisConfig>
#>  .. @ range_min     : num NA
#>  .. @ range_max     : num NA
#>  .. @ tick_count    : num NA
#>  .. @ tick_values   : NULL
#>  .. @ gridlines     : logi FALSE
#>  .. @ gridline_style: chr "dotted"
#>  .. @ ci_clip_factor: num 2
#>  .. @ include_null  : logi TRUE
#>  .. @ symmetric     : NULL
#>  .. @ null_tick     : logi TRUE
#>  .. @ marker_margin : logi TRUE
#>  @ layout       : <webforest::LayoutConfig>
#>  .. @ plot_position          : chr "right"
#>  .. @ table_width            : chr "auto"
#>  .. @ plot_width             : chr "auto"
#>  .. @ row_border             : logi TRUE
#>  .. @ row_border_style       : chr "solid"
#>  .. @ container_border       : logi FALSE
#>  .. @ container_border_radius: num 8
#>  @ group_headers: <webforest::GroupHeaderStyles>
#>  .. @ level1_font_size    : chr "0.9375rem"
#>  .. @ level1_font_weight  : num 600
#>  .. @ level1_italic       : logi FALSE
#>  .. @ level1_background   : NULL
#>  .. @ level1_border_bottom: logi FALSE
#>  .. @ level2_font_size    : chr "0.875rem"
#>  .. @ level2_font_weight  : num 500
#>  .. @ level2_italic       : logi FALSE
#>  .. @ level2_background   : NULL
#>  .. @ level2_border_bottom: logi FALSE
#>  .. @ level3_font_size    : chr "0.875rem"
#>  .. @ level3_font_weight  : num 400
#>  .. @ level3_italic       : logi FALSE
#>  .. @ level3_background   : NULL
#>  .. @ level3_border_bottom: logi FALSE
#>  .. @ indent_per_level    : num 16